refactor(storage): Move day file handling to explicitly deprecated file
