+++ internal/storage/old_day_file_handler.go
@@ -1,4 +1,4 @@
-package filehandling
+package storage
 
 import (
 	"bufio"
